Rename decorator module filenames to use kebab-case.
