+++ packages/labs/src/tests/frameworks/react/create-component_test.ts
@@ -14,7 +14,7 @@
 
 import {ReactiveElement} from '@lit/reactive-element';
 import {property} from '@lit/reactive-element/decorators/property.js';
-import {customElement} from '@lit/reactive-element/decorators/customElement.js';
+import {customElement} from '@lit/reactive-element/decorators/custom-element.js';
 import * as ReactModule from 'react';
 import 'react/umd/react.development.js';
 import 'react-dom/umd/react-dom.development.js';
