+++ packages/lit-element/rollup.config.js
@@ -22,14 +22,14 @@ export default litProdConfig({
     'hydrate-support',
     'private-ssr-support',
     'decorators',
-    'decorators/customElement',
-    'decorators/eventOptions',
+    'decorators/custom-element',
+    'decorators/event-options',
     'decorators/state',
     'decorators/property',
     'decorators/query',
-    'decorators/queryAll',
-    'decorators/queryAssignedNodes',
-    'decorators/queryAsync',
+    'decorators/query-all',
+    'decorators/query-assigned-nodes',
+    'decorators/query-async',
   ],
   external: ['lit-html', '@lit/reactive-element'],
   bundled: [
