+++ packages/lit-element/src/decorators.ts
@@ -19,11 +19,11 @@
  * not an arrow function.
  */
 
-export * from '@lit/reactive-element/decorators/customElement.js';
+export * from '@lit/reactive-element/decorators/custom-element.js';
 export * from '@lit/reactive-element/decorators/property.js';
 export * from '@lit/reactive-element/decorators/state.js';
-export * from '@lit/reactive-element/decorators/eventOptions.js';
+export * from '@lit/reactive-element/decorators/event-options.js';
 export * from '@lit/reactive-element/decorators/query.js';
-export * from '@lit/reactive-element/decorators/queryAll.js';
-export * from '@lit/reactive-element/decorators/queryAsync.js';
-export * from '@lit/reactive-element/decorators/queryAssignedNodes.js';
+export * from '@lit/reactive-element/decorators/query-all.js';
+export * from '@lit/reactive-element/decorators/query-async.js';
+export * from '@lit/reactive-element/decorators/query-assigned-nodes.js';
