+++ packages/lit-element/src/decorators/custom-element.ts
@@ -12,4 +12,4 @@
  * http://polymer.github.io/PATENTS.txt
  */
 
-export * from '@lit/reactive-element/decorators/queryAll.js';
+export * from '@lit/reactive-element/decorators/custom-element.js';
