+++ packages/lit-element/src/decorators/query-all.ts
@@ -12,4 +12,4 @@
  * http://polymer.github.io/PATENTS.txt
  */
 
-export * from '@lit/reactive-element/decorators/eventOptions.js';
+export * from '@lit/reactive-element/decorators/query-all.js';
