+++ packages/lit/rollup.config.js
@@ -18,13 +18,13 @@ export default litProdConfig({
   classPropertyPrefix: 'ϖ',
   entryPoints: [
     'decorators',
-    'decorators/customElement',
-    'decorators/eventOptions',
+    'decorators/custom-element',
+    'decorators/event-options',
     'decorators/property',
     'decorators/query',
-    'decorators/queryAll',
-    'decorators/queryAssignedNodes',
-    'decorators/queryAsync',
+    'decorators/query-all',
+    'decorators/query-assigned-nodes',
+    'decorators/query-async',
     'decorators/state',
     'directive-helpers',
     'directive',
