+++ packages/lit/src/decorators/event-options.ts
@@ -12,4 +12,4 @@
  * http://polymer.github.io/PATENTS.txt
  */
 
-export * from '@lit/reactive-element/decorators/queryAsync.js';
+export * from '@lit/reactive-element/decorators/event-options.js';
