+++ packages/lit/src/decorators/query-assigned-nodes.ts
@@ -12,4 +12,4 @@
  * http://polymer.github.io/PATENTS.txt
  */
 
-export * from '@lit/reactive-element/decorators/queryAssignedNodes.js';
+export * from '@lit/reactive-element/decorators/query-assigned-nodes.js';
