+++ packages/reactive-element/rollup.config.js
@@ -22,14 +22,14 @@ export default litProdConfig({
     'css-tag',
     'decorators',
     'decorators/base',
-    'decorators/customElement',
-    'decorators/eventOptions',
+    'decorators/custom-element',
+    'decorators/event-options',
     'decorators/state',
     'decorators/property',
     'decorators/query',
-    'decorators/queryAll',
-    'decorators/queryAssignedNodes',
-    'decorators/queryAsync',
+    'decorators/query-all',
+    'decorators/query-assigned-nodes',
+    'decorators/query-async',
   ],
   external: [],
   bundled: [
