+++ packages/reactive-element/src/decorators/custom-element.ts
