+++ packages/reactive-element/src/decorators/event-options.ts
