+++ packages/reactive-element/src/test/decorators/customElement_test.ts
@@ -12,7 +12,7 @@
  * http://polymer.github.io/PATENTS.txt
  */
 
-import {customElement} from '../../decorators/customElement.js';
+import {customElement} from '../../decorators/custom-element.js';
 import {generateElementName} from '../test-helpers.js';
 import {assert} from '@esm-bundle/chai';
